# api/openapi.yaml
openapi: 3.0.3
info:
  title: OVN Control Platform API
  description: |
    The OVN Control Platform provides a RESTful API for managing Open Virtual Network (OVN) resources.
    
    ## Features
    - Manage logical switches, routers, and ports
    - Configure ACLs and load balancers
    - Execute atomic transactions
    - Monitor resource health and metrics
    
    ## Authentication
    All API endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited. Check the following headers in responses:
    - `X-RateLimit-Limit`: Maximum requests per second
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
  version: 1.0.0
  contact:
    name: OVN Control Platform Team
    email: support@ovncp.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.ovncp.example.com/api/v1
    description: Production server
  - url: https://staging-api.ovncp.example.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /switches:
    get:
      tags:
        - Logical Switches
      summary: List all logical switches
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: name
          in: query
          schema:
            type: string
          description: Filter by switch name (supports wildcards)
      responses:
        '200':
          description: List of logical switches
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogicalSwitch'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags:
        - Logical Switches
      summary: Create a new logical switch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLogicalSwitch'
      responses:
        '201':
          description: Logical switch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalSwitch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /switches/{switchId}:
    get:
      tags:
        - Logical Switches
      summary: Get a logical switch by ID
      parameters:
        - $ref: '#/components/parameters/SwitchId'
      responses:
        '200':
          description: Logical switch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalSwitch'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Logical Switches
      summary: Update a logical switch
      parameters:
        - $ref: '#/components/parameters/SwitchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLogicalSwitch'
      responses:
        '200':
          description: Logical switch updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalSwitch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Logical Switches
      summary: Delete a logical switch
      parameters:
        - $ref: '#/components/parameters/SwitchId'
      responses:
        '204':
          description: Logical switch deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /routers:
    get:
      tags:
        - Logical Routers
      summary: List all logical routers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of logical routers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogicalRouter'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags:
        - Logical Routers
      summary: Create a new logical router
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLogicalRouter'
      responses:
        '201':
          description: Logical router created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogicalRouter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SwitchId:
      name: switchId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Logical switch UUID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        pattern: '^[a-zA-Z_]+:(asc|desc)$'
        default: 'created_at:desc'
      description: Sort field and order (e.g., name:asc)
  schemas:
...
    LogicalSwitch:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        subnet:
          type: string
          format: ipv4-cidr
        dns_servers:
          type: array
          items:
            type: string
            format: ipv4
        external_ids:
          type: object
          additionalProperties:
            type: string
        ports_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateLogicalSwitch:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 63
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9-]*[a-zA-Z0-9])?$'
        description:
          type: string
          maxLength: 255
        subnet:
          type: string
          format: ipv4-cidr
        dns_servers:
          type: array
          items:
            type: string
            format: ipv4
        external_ids:
          type: object
          additionalProperties:
            type: string
    UpdateLogicalSwitch:
      type: object
      properties:
        description:
          type: string
          maxLength: 255
        dns_servers:
          type: array
          items:
            type: string
            format: ipv4
        external_ids:
          type: object
          additionalProperties:
            type: string
    
    # Logical Router schemas
    LogicalRouter:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        external_gateway:
          type: string
          format: ipv4
        static_routes:
          type: array
          items:
            $ref: '#/components/schemas/StaticRoute'
        nat_rules:
          type: array
          items:
            $ref: '#/components/schemas/NATRule'
        external_ids:
          type: object
          additionalProperties:
            type: string
        ports_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateLogicalRouter:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 63
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9-]*[a-zA-Z0-9])?$'
        description:
          type: string
          maxLength: 255
        external_gateway:
          type: string
          format: ipv4
        static_routes:
          type: array
          items:
            $ref: '#/components/schemas/StaticRoute'
        nat_rules:
          type: array
          items:
            $ref: '#/components/schemas/NATRule'
        external_ids:
          type: object
          additionalProperties:
            type: string
...
    StaticRoute:
      type: object
      required:
        - destination
        - next_hop
      properties:
        destination:
          type: string
          format: ipv4-cidr
        next_hop:
          type: string
          format: ipv4
        metric:
          type: integer
          minimum: 0
          maximum: 65535
          default: 100
    NATRule:
      type: object
      required:
        - type
        - external_ip
        - internal_ip
      properties:
        type:
          type: string
          enum: [snat, dnat]
        external_ip:
          type: string
          format: ipv4
        internal_ip:
          type: string
          format: ipv4
        external_port:
          type: integer
          minimum: 1
          maximum: 65535
        internal_port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          type: string
          enum: [tcp, udp]
    
    # Logical Port schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: validation_error
            message: Invalid request parameters
            details:
              field: name
              reason: Name must be alphanumeric
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
            message: Insufficient permissions for this operation
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: Resource not found
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: conflict
            message: Resource with this name already exists
...